test_dep4 ref in test_mc1step for improved hessian
test_dep4 and test_dep4_df in test_mc1step.py check the nonconverged
energy after one macrocycle. Because an error in the Hessian was
fixed, this reference had to change. It is now *lower* than it was
before, implying an *improved* hessian-vector product which should
generally converge orbitals faster.