Undo commit 7a3531c "error fix", which was wrong
Dunno if I'm gonna even try to merge this now, but for the record
the error that I thought I fixed was actually an error in the
seminumerical hessian-vector product I was using as a reference.
This takes a lot of math to figure out. However the fix related
to issue 701 is legitimate and maybe these expanded tests are still
useful?